fynd/requirements.txt
Django==3.1
django-filter==2.4.0
django-jsonfield==1.4.1
djangorestframework==3.12.2
requests==2.22.0
textblob==0.15.3
python==3.8.5
pip==20.0.2